sandbox: rename WASI interfaces to wasi*
